+++ examples/fully_async_async_std.rs
@@ -16,6 +16,7 @@ struct MyTask {
 impl AsyncTask for MyTask {
     // Different tasks can return different results. If successful, this task returns an `f64`.
     type Output = f64;
+    type Affinity = DispatchAny;
 
     // Include the custom code MyTask needs.
     async fn register<'frame>(mut frame: AsyncGcFrame<'frame>) -> JlrsResult<()> {
@@ -76,7 +77,10 @@ async fn main() {
     {
         // Include the custom code MyTask needs by registering it.
         let (sender, receiver) = async_std::channel::bounded(1);
-        julia.register_task::<MyTask, _>(sender).await;
+        julia
+            .register_task::<MyTask, _>(sender)
+            .dispatch_any()
+            .await;
         receiver.recv().await.unwrap().unwrap();
     }
 
@@ -92,6 +96,7 @@ async fn main() {
             },
             sender1,
         )
+        .dispatch_any()
         .await;
 
     julia
@@ -102,6 +107,7 @@ async fn main() {
             },
             sender2,
         )
+        .dispatch_any()
         .await;
 
     // Receive the results of the tasks.